# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  handler.BacktestRequest:
    properties:
      commission:
        example: 0.001
        type: number
      end_date:
        example: "2025-10-17"
        type: string
      fast_period:
        description: Strategy parameters (MA Cross)
        example: 10
        type: integer
      initial_balance:
        example: 10000
        type: number
      interval:
        example: 1h
        type: string
      position_size:
        description: Position size
        example: 0.01
        type: number
      rsi_overbought:
        example: 70
        type: number
      rsi_oversold:
        example: 30
        type: number
      rsi_period:
        description: Strategy parameters (RSI)
        example: 14
        type: integer
      slow_period:
        example: 30
        type: integer
      start_date:
        example: "2025-07-01"
        type: string
      strategy:
        example: ma_cross
        type: string
      symbol:
        example: BTCUSDT
        type: string
    required:
    - end_date
    - interval
    - start_date
    - symbol
    type: object
  handler.BacktestResponse:
    properties:
      candles_used:
        example: 2152
        type: integer
      commission:
        example: 0.001
        type: number
      duration:
        example: 2151h0m0s
        type: string
      end_time:
        example: "2025-10-17T08:00:00Z"
        type: string
      final_equity:
        example: 9866.65
        type: number
      initial_balance:
        description: Financial metrics
        example: 10000
        type: number
      interval:
        example: 1h
        type: string
      losing_trades:
        example: 29
        type: integer
      max_drawdown:
        example: 206.59
        type: number
      max_drawdown_pct:
        example: 2.07
        type: number
      recent_trades:
        description: Trade details (optional, limited to first 20)
        items:
          $ref: '#/definitions/handler.TradeDetail'
        type: array
      sharpe_ratio:
        description: Risk metrics
        example: -2.16
        type: number
      start_time:
        description: Time metrics
        example: "2025-07-19T17:00:00Z"
        type: string
      strategy:
        description: Configuration
        example: MA_Cross_10_30
        type: string
      symbol:
        example: BTCUSDT
        type: string
      total_return:
        example: -1.33
        type: number
      total_return_pct:
        example: -1.33%
        type: string
      total_trades:
        description: Trade statistics
        example: 82
        type: integer
      win_rate:
        example: 29.27
        type: number
      win_rate_pct:
        example: 29.27%
        type: string
      winning_trades:
        example: 12
        type: integer
    type: object
...
  handler.TradeDetail:
    properties:
      balance:
        example: 9328.59
        type: number
      fee:
        example: 1.19
        type: number
      position:
        example: 0.01
        type: number
      price:
        example: 119362.67
        type: number
      quantity:
        example: 0.01
        type: number
      reason:
        example: 'Golden Cross: Fast MA (118128.46) > Slow MA (118117.20)'
        type: string
      side:
        example: BUY
        type: string
      timestamp:
        example: "2025-07-21T16:00:00Z"
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Cryptocurrency quantitative trading API server
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Crypto Quant API
  version: "1.0"
paths:
  /backtest/run:
    post:
      consumes:
      - application/json
      description: Execute a backtest with specified parameters and strategy
      parameters:
      - description: Backtest parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.BacktestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.BacktestResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Run a backtest
      tags:
      - backtest
...
swagger: "2.0"